Remove .gitignore and update output file formats
The .gitignore file has been removed and the output formats in RunOptimiser.py were updated. The removal of the .gitignore file simplifies the project structure and potentially facilitates a better interaction with the Git version control system. Updates in RunOptimiser.py include adjusting the output file formats and locations for better organization and readability.